ilyad says this gets rid of a warning on 32bit systems

+++ sketch.c
@@ -256,7 +256,7 @@ void register_builtin(char *name, builtin_t func) {
   uint64_t value = T_FUNC | BLTIN_MASK;
   uint32_t index = next_cell;
   cells[next_cell++] = value;
-  cells[next_cell++] = (uint64_t)func;
+  cells[next_cell++] = (uint64_t)(uintptr_t)func;
   set_symbol(name, strlen(name), index);
 }
